feat(net): move Display of PacketStats
